requirements.txt
autopep8==1.6.0
blis==0.7.7
catalogue==2.0.7
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
cymem==2.0.6
docx2txt==0.8
docxcompose==1.3.4
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.3.0/en_core_web_lg-3.3.0-py3-none-any.whl
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.3.0/en_core_web_md-3.3.0-py3-none-any.whl
en-core-web-trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.3.0/en_core_web_trf-3.3.0-py3-none-any.whl
et-xmlfile==1.1.0
filelock==3.7.0
flake8==4.0.1
huggingface-hub==0.7.0
idna==3.3
inflect==5.6.0
Jinja2==3.1.2
joblib==1.1.0
langcodes==3.3.0
lxml==4.8.0
MarkupSafe==2.1.1
mccabe==0.6.1
murmurhash==1.0.7
nltk==3.7
numpy==1.22.4
openpyxl==3.0.10
packaging==21.3
pathy==0.6.1
preshed==3.0.6
pycodestyle==2.8.0
pydantic==1.8.2
pyflakes==2.4.0
pyparsing==3.0.9
python-docx==0.8.11
python-dotenv==0.20.0
PyYAML==6.0
regex==2022.4.24
requests==2.27.1
sacremoses==0.0.53
six==1.16.0
smart-open==5.2.1
spacy==3.3.0
spacy-alignments==0.8.5
spacy-legacy==3.0.9
spacy-loggers==1.0.2
spacy-transformers==1.1.5
srsly==2.4.3
target-matcher==0.0.3
thinc==8.0.16
tika==1.24
tokenizers==0.12.1
toml==0.10.2
torch==1.11.0
tqdm==4.64.0
transformers==4.17.0
typer==0.4.1
typing_extensions==4.2.0
urllib3==1.26.9
wasabi==0.9.1
